requirements.txt
pyautogui
pynput
opencv-python
pillow
pytesseract
requests
pandas
loguru
pytest
schedule
python-dotenv